+++ crates/prek/src/cli/run/mod.rs
@@ -3,6 +3,7 @@ pub(crate) use filter::{
     collect_run_input,
 };
 pub(crate) use install::{InstallCache, install_hooks};
+pub(crate) use reporter::{HookRunReporter, project_status_marker};
 pub(crate) use run::run;
 pub(crate) use selector::{SelectorSource, Selectors};
 
@@ -10,6 +11,7 @@ mod diff;
 mod filter;
 mod install;
 mod keeper;
+mod reporter;
 #[allow(clippy::module_inception)]
 mod run;
 mod selector;
