+++ crates/prek/src/languages/dart.rs
@@ -22,7 +22,8 @@ use prek_consts::prepend_paths;
 use serde::{Deserialize, Serialize};
 use tracing::debug;
 
-use crate::cli::reporter::{HookInstallReporter, HookRunReporter};
+use crate::cli::reporter::HookInstallReporter;
+use crate::cli::run::HookRunReporter;
 use crate::hook::{Hook, InstallInfo, InstalledHook};
 use crate::languages::LanguageImpl;
 use crate::process::Cmd;
