+++ crates/prek/src/languages/fail.rs
@@ -4,7 +4,8 @@ use std::sync::Arc;
 
 use anyhow::Result;
 
-use crate::cli::reporter::{HookInstallReporter, HookRunReporter};
+use crate::cli::reporter::HookInstallReporter;
+use crate::cli::run::HookRunReporter;
 use crate::hook::{Hook, InstallInfo, InstalledHook};
 use crate::languages::LanguageImpl;
 use crate::store::Store;
