+++ crates/prek/src/languages/golang/golang.rs
@@ -7,7 +7,8 @@ use anyhow::Context;
 use prek_consts::env_vars::EnvVars;
 use prek_consts::prepend_paths;
 
-use crate::cli::reporter::{HookInstallReporter, HookRunReporter};
+use crate::cli::reporter::HookInstallReporter;
+use crate::cli::run::HookRunReporter;
 use crate::hook::{Hook, InstallInfo, InstalledHook};
 use crate::languages::LanguageImpl;
 use crate::languages::golang::GoRequest;
