+++ crates/prek/src/languages/julia.rs
@@ -5,7 +5,8 @@ use std::sync::Arc;
 use anyhow::{Context, Result};
 use tracing::debug;
 
-use crate::cli::reporter::{HookInstallReporter, HookRunReporter};
+use crate::cli::reporter::HookInstallReporter;
+use crate::cli::run::HookRunReporter;
 use crate::hook::{Hook, InstallInfo, InstalledHook};
 use crate::languages::LanguageImpl;
 use crate::process::Cmd;
